+++ src/app/content/studyGuides/hooks/index.ts
@@ -1,3 +1,12 @@
+import { initializeMyHighlightsSummaryHook } from './initializeStudyGuides';
+import { loadMoreHook } from './loadMore';
 import loadStudyGuides from './locationChange';
 
-export { loadStudyGuides };
+export {
+  loadStudyGuides,
+};
+
+export default [
+  initializeMyHighlightsSummaryHook,
+  loadMoreHook,
+]
