+++ src/app/content/studyGuides/hooks/initializeStudyGuides.ts
@@ -1,0 +1,34 @@
+import { GetHighlightsSummarySetsEnum, GetHighlightsSummarySourceTypeEnum } from '@openstax/highlighter/dist/api';
+import { ActionHookBody } from '../../../types';
+import { actionHook, assertDefined } from '../../../utils';
+import { summaryPageSize } from '../../constants';
+import { bookAndPage } from '../../selectors';
+import { extractTotalCounts } from '../../utils/sharedHighlightsUtils';
+import { openStudyGuides, receiveStudyGuidesSummary, receiveStudyGuidesTotalCounts } from '../actions';
+import * as select from '../selectors'
+import { loadMoreStudyGuidesHighlights } from './loadMore';
+
+export const hookBody: ActionHookBody<typeof openStudyGuides> = (services) => async() => {
+  const state = services.getState();
+
+  const {book} = bookAndPage(state);
+  const needsInitialization = () => select.studyGuidesEnabled(state)
+    && select.studyGuidesSummaryHighlights(state) === null
+    && select.studyGuidesIsLoading(state) === false;
+
+  if (!book || !needsInitialization()) { return; }
+
+  const studyGuidesSummary = await services.highlightClient.getHighlightsSummary({
+    scopeId: book.id,
+    sets: [GetHighlightsSummarySetsEnum.Curatedopenstax],
+    sourceType: GetHighlightsSummarySourceTypeEnum.OpenstaxPage,
+  });
+
+  const countsPerSource = assertDefined(studyGuidesSummary.countsPerSource, 'summary response is invalid');
+  services.dispatch(receiveStudyGuidesTotalCounts(extractTotalCounts(countsPerSource)));
+
+  const {formattedHighlights, pagination} = await loadMoreStudyGuidesHighlights(services, summaryPageSize);
+  services.dispatch(receiveStudyGuidesSummary(formattedHighlights, pagination));
+};
+
+export const initializeMyHighlightsSummaryHook = actionHook(openStudyGuides, hookBody);
